fixing quadtree

+++ src/quadtree.c
@@ -14,38 +14,40 @@ static void split_node(struct quad_tree *tree, struct quad_node *node) {
         return;
     }
 
-    void *children_data[4] = {NULL};
-    tree->split(tree, node->data, children_data);
-
     for (size_t i = 0; i < 4; ++i) {
-        struct rect bounds = rect_subdivide_i(node->bounds, i);
+        struct rect bounds = rect_subdivide_i(node->bounds, i, 4);
         node->children[i] = malloc(sizeof(struct quad_node));
-        quad_node_init(
-                node->children[i], node->depth + 1, children_data[i], bounds);
+
+        void *data = tree->split_i(tree, node->data, bounds);
+
+        quad_node_init(node->children[i], node->depth + 1, data, bounds);
 
         if (tree->should_split(tree, node->children[i])) {
             split_node(tree, node->children[i]);
         }
+    }
+
+    if (node->data != NULL && tree->destroy_data != NULL) {
+        tree->destroy_data(node->data);
     }
 }
 
 void quad_tree_create(size_t max_depth,
         size_t max_objects,
         bool (*should_split)(struct quad_tree *, struct quad_node *),
-        void (*split)(struct quad_tree *, void *, void **),
+        void *(*split)(struct quad_tree *, void *,  struct rect bounds),
         void (*destroy_data)(void *),
         void *node_data,
         struct rect root_bounds,
         struct quad_tree *tree_out) {
     assert(should_split != NULL);
     assert(split != NULL);
-    assert(destroy_data != NULL);
     assert(tree_out != NULL);
 
     tree_out->max_depth = max_depth;
     tree_out->max_objects = max_objects;
     tree_out->should_split = should_split;
-    tree_out->split = split;
+    tree_out->split_i = split;
     tree_out->destroy_data = destroy_data;
 
     tree_out->root = malloc(sizeof(struct quad_node));
@@ -68,8 +70,6 @@ void quad_node_init(
 }
 
 static void destroy_node(struct quad_node *node, void (*destroy_data)(void *)) {
-    assert(destroy_data != NULL);
-
     if (node == NULL) {
         return;
     }
@@ -78,14 +78,41 @@ static void destroy_node(struct quad_node *node, void (*destroy_data)(void *)) {
         destroy_node(node->children[i], destroy_data);
     }
 
-    if (node->data != NULL) {
+    if (node->data != NULL && destroy_data != NULL) {
         destroy_data(node->data);
     }
 
     free(node);
 }
 
-void quadtree_destroy(struct quad_tree *tree) {
+void *quad_tree_query(struct quad_tree *tree, vec3 position) {
+    assert(tree != NULL);
+
+    struct quad_node *current = tree->root;
+
+    while (current != NULL) {
+        struct quad_node *next = NULL;
+
+        for (size_t i = 0; i < 4; ++i) {
+            if (current->children[i] != NULL
+                    && position_within_rect(position, current->children[i]->bounds)) {
+                next = current->children[i];
+                break;
+            }
+        }
+
+        if (next == NULL) {
+            return current->data;
+        }
+
+        current = next;
+    }
+
+    return NULL;
+}
+
+
+void quad_tree_destroy(struct quad_tree *tree) {
     assert(tree != NULL);
 
     destroy_node(tree->root, tree->destroy_data);
